NX-34: close deltas packet marker byte sets
Replace raw marker bytes on source and native reference-marker and
type-150 packets with shared closed enums. The binary parsers construct
the enums directly and no longer retain separate literal admission checks.
Serde keeps the numeric wire values and rejects values outside each set.
Existing tests retain their byte expectations at the numeric boundary;
new coverage checks all 256 inputs for both enums.

Packet reference and numeric payload invariants and the wider NX-34 audit
remain open; this commit closes only these two marker fields.

Validation: cargo check -q -p cadmpeg-codec-nx --tests.
Hook-Bypass-Reason: user requires scoped compile checks between items and the full gate once at completion.

+++ crates/cadmpeg-codec-nx/src/deltas/tests/reference_and_tombstone_packets.rs
@@ -141,7 +141,7 @@ fn deltas_reference_marker_packets_decode_extended_references_atomically() {
 
     assert_eq!(census.reference_marker_packets.len(), 1);
     assert_eq!(census.reference_marker_packets[0].reference, 40_000);
-    assert_eq!(census.reference_marker_packets[0].marker, 0x56);
+    assert_eq!(u8::from(census.reference_marker_packets[0].marker), 0x56);
     assert_eq!(census.reference_marker_packets[0].offset, 0);
     assert_eq!(census.reference_marker_packets[0].end, packet.len());
     assert_eq!(census.bytes_decoded, packet.len());
